this is just a reorg of the previous code
webrtc_server.py was getting too big and audio functions and member vars felt like a good candidate for decomposition.  Sound output on the bot is still slow and low.
