+++ src/commons/webrtc_server.py
@@ -8,19 +8,16 @@ between the webapp and onboard_ui service for browser-to-display video streaming
 import asyncio
 import json
 import logging
-import queue
-import threading
 from typing import Optional
 
 import aiohttp
-import numpy as np
-import sounddevice as sd
 from aiohttp import web, WSMsgType
 from aiohttp_cors import setup as cors_setup, ResourceOptions
 from aiortc import RTCPeerConnection, RTCSessionDescription
 
 from basic_bot.commons import log
 from commons.constants import D2_OUI_WEBRTC_HOST, D2_OUI_WEBRTC_PORT
+from commons.audio_stream_player import AudioStreamPlayer
 
 logger = logging.getLogger(__name__)
 
@@ -37,11 +34,7 @@ class WebRTCSignalingServer:
         self.video_callback = video_callback
         self.peer_connection: Optional[RTCPeerConnection] = None
         self.websocket: Optional[aiohttp.ClientWebSocketResponse] = None
-
-        # Audio playback setup
-        self.audio_queue = queue.Queue(maxsize=100)  # Buffer for audio frames
-        self.audio_stream = None
-        self.audio_thread = None
+        self.audio_player = AudioStreamPlayer()
 
         # Setup CORS for cross-origin requests from webapp
         cors = cors_setup(
@@ -92,7 +85,7 @@ class WebRTCSignalingServer:
         self.websocket = None
 
         # Clean up audio stream
-        self._cleanup_audio_stream()
+        self.audio_player.cleanup_audio_stream()
 
         # Clean up peer connection
         if self.peer_connection:
@@ -193,119 +186,20 @@ class WebRTCSignalingServer:
 
             # Start audio stream on first audio frame
             first_frame = await track.recv()
-            await self._setup_audio_stream(first_frame)
+            await self.audio_player.setup_audio_stream(first_frame)
 
             # Queue the first frame
-            self._queue_audio_frame(first_frame)
+            self.audio_player.queue_audio_frame(first_frame)
 
             # Process remaining frames
             while True:
                 frame = await track.recv()
-                self._queue_audio_frame(frame)
+                self.audio_player.queue_audio_frame(frame)
 
         except Exception as e:
             log.error(f"Error processing audio track: {e}")
         finally:
-            self._cleanup_audio_stream()
-
-    async def _setup_audio_stream(self, first_frame):
-        """Setup sounddevice audio stream based on first audio frame."""
-        try:
-            # Convert frame to numpy array to get audio properties
-            audio_data = np.frombuffer(
-                first_frame.to_ndarray().tobytes(), dtype=np.int16
-            )
-
-            # Get audio properties from the frame
-            sample_rate = first_frame.sample_rate
-            channels = (
-                len(first_frame.layout.channels)
-                if hasattr(first_frame, "layout")
-                else 1
-            )
-
-            log.info(f"Setting up audio stream: {sample_rate}Hz, {channels} channels")
-
-            # Create audio stream with callback
-            self.audio_stream = sd.OutputStream(
-                samplerate=sample_rate,
-                channels=channels,
-                dtype=np.int16,
-                callback=self._audio_callback,
-                blocksize=1024,  # Small buffer for low latency
-                latency="low",
-            )
-
-            self.audio_stream.start()
-            log.info("Audio stream started successfully")
-
-        except Exception as e:
-            log.error(f"Error setting up audio stream: {e}")
-
-    def _audio_callback(self, outdata, frames, time, status):
-        """Callback function for sounddevice audio stream."""
-        if status:
-            log.warning(f"Audio callback status: {status}")
-
-        try:
-            # Get audio data from queue
-            audio_data = self.audio_queue.get_nowait()
-
-            # Ensure data fits in output buffer
-            if len(audio_data) <= len(outdata):
-                outdata[: len(audio_data)] = audio_data.reshape(-1, 1)
-                # Pad with zeros if needed
-                if len(audio_data) < len(outdata):
-                    outdata[len(audio_data) :] = 0
-            else:
-                # Truncate if too long
-                outdata[:] = audio_data[: len(outdata)].reshape(-1, 1)
-
-        except queue.Empty:
-            # No audio data available, output silence
-            outdata.fill(0)
-        except Exception as e:
-            log.error(f"Error in audio callback: {e}")
-            outdata.fill(0)
-
-    def _queue_audio_frame(self, frame):
-        """Convert and queue audio frame for playback."""
-        try:
-            # Convert WebRTC frame to numpy array
-            audio_data = np.frombuffer(frame.to_ndarray().tobytes(), dtype=np.int16)
-
-            # Add to queue (drop old frames if queue is full)
-            try:
-                self.audio_queue.put_nowait(audio_data)
-            except queue.Full:
-                # Remove oldest frame and add new one
-                try:
-                    self.audio_queue.get_nowait()
-                    self.audio_queue.put_nowait(audio_data)
-                except queue.Empty:
-                    pass
-
-        except Exception as e:
-            log.error(f"Error queuing audio frame: {e}")
-
-    def _cleanup_audio_stream(self):
-        """Clean up audio stream resources."""
-        try:
-            if self.audio_stream:
-                self.audio_stream.stop()
-                self.audio_stream.close()
-                self.audio_stream = None
-                log.info("Audio stream cleaned up")
-
-            # Clear audio queue
-            while not self.audio_queue.empty():
-                try:
-                    self.audio_queue.get_nowait()
-                except queue.Empty:
-                    break
-
-        except Exception as e:
-            log.error(f"Error cleaning up audio stream: {e}")
+            self.audio_player.cleanup_audio_stream()
 
     async def send_message(self, message):
         """Send message to WebSocket client."""
@@ -336,7 +230,7 @@ class WebRTCSignalingServer:
     async def stop_server(self, runner):
         """Stop the WebRTC signaling server."""
         # Clean up audio stream
-        self._cleanup_audio_stream()
+        self.audio_player.cleanup_audio_stream()
 
         if self.peer_connection:
             await self.peer_connection.close()
